更改图标，电量信息区分MK114和MK115

+++ app/src/main/java/com/moko/lifex/entity/PowerInfo.java
@@ -1,7 +1,7 @@
 package com.moko.lifex.entity;
 
 public class PowerInfo {
-    public int voltage;
-    public int current;
-    public int power;
+    public float voltage;
+    public float current;
+    public float power;
 }
